+++ src/setup.jl
@@ -57,8 +57,8 @@ export read_amsr
 export read_argo
 export read_argo_index
 export read_ctd_cnv
+export read_ctd_exchange
 export read_ctd_rsk
-export read_ctd_woce
 export read_section
 export read_topography
 export salinity_from_conductivity